+++ src/main/kotlin/dev/nyon/magnetic/Main.kt
@@ -1,0 +1,10 @@
+@file:Suppress("unused")
+
+package dev.nyon.magnetic
+
+//? if <=1.20.6
+/*lateinit var magnetic: MagneticEnchantment*/
+
+fun init() {
+    DropEvent
+}