+++ src/main/kotlin/dev/nyon/magnetic/config/ModMenuImpl.kt
@@ -1,0 +1,13 @@
+package dev.nyon.magnetic.config
+
+import com.terraformersmc.modmenu.api.ConfigScreenFactory
+import com.terraformersmc.modmenu.api.ModMenuApi
+
+@Suppress("unused")
+class ModMenuImpl : ModMenuApi {
+    override fun getModConfigScreenFactory(): ConfigScreenFactory<*> {
+        return ConfigScreenFactory {
+            generateConfigScreen(it)
+        }
+    }
+}